feat: add support for parsing as composite classes as default
Before this commit, the TiledProjectParser could only parse composite classes with the `@composite:` prefix, which may become tedious for users whose use case involves mainly composition.

Let's add a parser options optional argument to the parse method of the TiledProjectParser, which allows it to parse the project with types of nested properties by default, similar to how Tiled displays it in the GUI. Only names with the `@inherit` prefix will be flattened. This allows user to choose their parsing method based on their own use case.